--- learning_strategies/__init__.py
@@ -1,3 +1,0 @@
-from .gaussian import Gaussian
-
-all = ["Gaussian"]
